Fail silently for Net::HTTP exceptions
Some of the most common exceptions Net::HTTP
requests raise are Timeout::Error, Errno::ECONNRESET
and Net::HTTPBadResponse. Those are now ignored and
do not prevent signing up.

The default Net::HTTP timeout setting for open
and read is 60 seconds, which is too long to block
a sign up for, so more reasonable defaults are set
with the option to override them.