Fix “no implicit conversion of nil into String”
If the devise model is :trackable, the model will be updated and saved at which point validations are run. However password is cleared on authentication so `Digest::SHA1.hexdigest(password).upcase` fails because password is nil. Add a guard clause to prevent the validation from running if the password isn’t present.